Consolidate example code
