+++ example/src/home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Connect, query, mutation } from '../src/index';
+import { Connect, query, mutation } from '../../src/index';
 import TodoList from './todo-list';
 import TodoForm from './todo-form';
 import Loading from './loading';
